requirements.txt
Django==3.0.3
Markdown==3.2.1
mysql-connector-python==8.0.19
mysqlclient==1.4.6
PyYAML==5.3
django-pdb==0.6.2
django-rest-swagger==2.2.0
django-reversion==3.0.7
django-suit==0.2.28
djangorestframework==3.11.0
requests==2.23.0
six==1.14.0
django-environ==0.4.5
dj-static==0.0.6
kubernetes==10.0.1
werkzeug==1.0.0
sphinx==2.4.3
sphinx-rtd-theme==0.4.3
recommonmark==0.6.0
pandas==1.0.1
